Classes API use Integer instead of Long
Changed all classes of User and Tournament to use integer instead of longs

+++ src/main/java/com/fontys/api/entities/Tournament.java
@@ -9,7 +9,7 @@ import javax.persistence.*;
 public class Tournament
 {
     @GeneratedValue @Id
-    private Long id;
+    private Integer id;
     private String name;
     private String description;
     @OneToOne
@@ -26,7 +26,7 @@ public class Tournament
         this.numberOfTeams = numberOfTeams;
     }
 
-    public Tournament(long id, String name, String description, User user, int numberOfTeams)
+    public Tournament(Integer id, String name, String description, User user, int numberOfTeams)
     {
         this.id = id;
         this.name = name;
