+++ src/main/java/com/fontys/api/repositories/TournamentRepository.java
@@ -3,6 +3,6 @@ package com.fontys.api.repositories;
 import com.fontys.api.entities.Tournament;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TournamentRepository extends JpaRepository<Tournament, Long>
+public interface TournamentRepository extends JpaRepository<Tournament, Integer>
 {
 }
