+++ src/test/java/com/fontys/api/mockrepositories/MockTournamentRepository.java
@@ -3,6 +3,6 @@ package com.fontys.api.mockrepositories;
 import com.fontys.api.entities.Tournament;
 import com.fontys.api.repositories.TournamentRepository;
 
-public class MockTournamentRepository extends MockJPARepository<Tournament, Long> implements TournamentRepository
+public class MockTournamentRepository extends MockJPARepository<Tournament, Integer> implements TournamentRepository
 {
 }
