+++ src/test/java/com/fontys/api/mockrepositories/MockUserRepository.java
@@ -3,6 +3,6 @@ package com.fontys.api.mockrepositories;
 import com.fontys.api.entities.User;
 import com.fontys.api.repositories.UserRepository;
 
-public class MockUserRepository extends MockJPARepository<User, Long> implements UserRepository
+public class MockUserRepository extends MockJPARepository<User, Integer> implements UserRepository
 {
 }
